Ironed out all the issues.  We have a winner!
An AI pawn now walks forward through a number of trigger boxes, both those being tested, and several that are themselves tests.

The test logs when health checks pass, and fails the test when a health check fails, ending the test.  The log message on failure is in the format "Expectd X, but found Y".

On the C++ side, we've had logging every time SetHealth is changed, whether the value changes or not.  And we log when we fire the Death Delegate of Death.  That's been in place since yesterday, but is mentioned here for completeness.